pyproject.toml
[tool.poetry]
name = "analytics-inventory-transfers-model"
version = "0.1.0"
description = ""
authors = ["fabiancastano <fabian.castano@frubana.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
numpy = "^2.1.1"
python-json-logger = "^2.0.7"
jupyter = "^1.1.1"
pandas = "^2.2.3"
scipy = "^1.14.1"
python-dotenv = "^1.0.1"
statsmodels = "^0.14.4"
highsbox = "^1.7.2.post2"
pulp = "^2.9.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
